Experiment 02 Phase 7: SHAP, PDP/ALE and symbolic regression
SHAP (random_forest, 2,000 rows explained): calidad_ponderada dominates at
0.181 mean |SHAP|, five times the next feature (comuna=Las Condes, 0.035).
The model is largely recovering the structure of the SII's own valuation
instrument, whose unit-construction-value tables are indexed by quality -
another reminder that this predicts assessed value, not market price.

PDP and ALE are plotted together for the top three continuous features,
because they diverge under correlated features and that divergence is
information, not noise.

Symbolic regression found a genuinely readable closed form (complexity 5):

    log10(UF/m2) = 0.6058 / calidad + 1.3136

Scored honestly, this is the surprise of the phase: on held-out manzanas the
two-term expression beats a random forest given the same rows and the same
numeric-only features (MAE 0.0380 vs 0.0745, 0.51x). With few rows and no
locational feature the forest overfits; the formula generalises.

Getting to that number took fixing the comparison twice, and both flattered
the forest:
  - first it was scored in-sample (MAE 0.0028 vs its honest 0.0394 in CV),
  - then the forest was still trained on the rows the expression was tested on.
The reference is now a forest fit on exactly PySR's training rows and feature
set, scored on the same held-out manzanas.

The formula's limits are recorded too: across quality levels it tracks the
ordering well (Pearson 0.92) but overpredicts both tails - 83 UF/m2 against a
real median of 56 at Superior, 27 against 13.6 at Inferior - compressing a
4.1x real spread into 3.1x.

Also fixes es_departamento to float: sklearn's partial_dependence rejects
integer columns, and the 0/1 values are identical either way.

+++ experiments/02-plusvalia/features.py
@@ -101,7 +101,9 @@ def build_features(df: pd.DataFrame, cfg: dict) -> tuple[pd.DataFrame, pd.Series
     # imputing 0 -> -inf would be nonsense, and 0 is meaningful, not missing.
     land = df["sup_terreno_m2"].replace(0, np.nan)
     f["log10_sup_terreno_m2"] = np.log10(land)
-    f["es_departamento"] = (df["sup_terreno_m2"] == 0).astype(int)
+    # float, not int: sklearn's partial_dependence refuses integer columns
+    # (implicit rounding), and the 0/1 values are identical either way.
+    f["es_departamento"] = (df["sup_terreno_m2"] == 0).astype(float)
 
     f["calidad_ponderada"] = df["calidad_ponderada"]
     f["anio_construccion"] = df["anio_construccion"]
